requirements.txt
aiogram==3.21.0
aiohttp>=3.12,<4
magic-filter>=1.0.12
httpx>=0.28
gspread
google-auth
google-auth-oauthlib
python-dotenv
